requirements.txt
fpdf2==2.7.6
pathspec==0.11.2
PyPDF2==3.0.1
Pygments==2.15.1
pyinstaller==5.13.0
pytest==7.3.1